moved all header files into proper directories

+++ include/arch/armv7/generic_timer.h
@@ -2,7 +2,7 @@
 #define __GENERIC_TIMER_H___
 
 /* Generic Timer */
-#define read_cntfrq()           ({ uint32_t rval; asm volatile(\
+#define read_cntfrq()           ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c0, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -10,7 +10,7 @@
                                 " mcr     p15, 0, %0, c14, c0, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cnthctl()          ({ uint32_t rval; asm volatile(\
+#define read_cnthctl()          ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 4, %0, c14, c1, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -18,7 +18,7 @@
                                 " mcr     p15, 4, %0, c14, c1, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cnthp_ctl()        ({ uint32_t rval; asm volatile(\
+#define read_cnthp_ctl()        ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 4, %0, c14, c2, 1\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -26,17 +26,17 @@
                                 " mcr     p15, 4, %0, c14, c2, 1\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cnthp_cval()       ({ uint32_t v1, v2; asm volatile(\
+#define read_cnthp_cval()       ({ unsigned int v1, v2; asm volatile(\
                                 " mrrc     p15, 6, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
+                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
 
 #define write_cnthp_cval(val) 	asm volatile(\
                               	  " mcrr     p15, 6, %0, %1, c14\n\t" \
 								  : : "r" ((val) & 0xFFFFFFFF), "r" ((val) >> 32) \
 								  : "memory", "cc")
 
-#define read_cnthp_tval()       ({ uint32_t rval; asm volatile(\
+#define read_cnthp_tval()       ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 4, %0, c14, c2, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -44,7 +44,7 @@
                                 " mcr     p15, 4, %0, c14, c2, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntkctl()          ({ uint32_t rval; asm volatile(\
+#define read_cntkctl()          ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c1, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -52,7 +52,7 @@
                                 " mcr     p15, 0, %0, c14, c1, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntp_ctl()         ({ uint32_t rval; asm volatile(\
+#define read_cntp_ctl()         ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c2, 1\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -60,17 +60,17 @@
                                 " mcr     p15, 0, %0, c14, c2, 1\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntp_cval()        ({ uint32_t v1, v2; asm volatile(\
+#define read_cntp_cval()        ({ unsigned int v1, v2; asm volatile(\
                                 " mrrc     p15, 2, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
+                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
 
 #define write_cntp_cval(val)  	asm volatile(\
                               	  " mcrr     p15, 2, %0, %1, c14\n\t" \
 								  : : "r" ((val) & 0xFFFFFFFF), "r" ((val) >> 32) \
 								  : "memory", "cc")
 
-#define read_cntp_tval()        ({ uint32_t rval; asm volatile(\
+#define read_cntp_tval()        ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c2, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -78,12 +78,12 @@
                                 " mcr     p15, 0, %0, c14, c2, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntpct()           ({ uint32_t v1, v2; asm volatile(\
+#define read_cntpct()           ({ unsigned int v1, v2; asm volatile(\
                                 " mrrc     p15, 0, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
+                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
 
-#define read_cntv_ctl()         ({ uint32_t rval; asm volatile(\
+#define read_cntv_ctl()         ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c3, 1\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -91,17 +91,17 @@
                                 " mcr     p15, 0, %0, c14, c3, 1\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntv_cval()        ({ uint32_t v1, v2; asm volatile(\
+#define read_cntv_cval()        ({ unsigned int v1, v2; asm volatile(\
                                 " mrrc     p15, 3, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
+                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
 
 #define write_cntv_cval(val)  asm volatile(\
                               " mcrr     p15, 3, %0, %1, c14\n\t" \
                               : : "r" ((val) & 0xFFFFFFFF), "r" ((val) >> 32) \
                               : "memory", "cc")
 
-#define read_cntv_tval()        ({ uint32_t rval; asm volatile(\
+#define read_cntv_tval()        ({ unsigned int rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c3, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -109,15 +109,15 @@
                                 " mcr     p15, 0, %0, c14, c3, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntvct()           ({ uint32_t v1, v2; asm volatile(\
+#define read_cntvct()           ({ unsigned int v1, v2; asm volatile(\
                                 " mrrc     p15, 1, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
+                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
 
-#define read_cntvoff()          ({ uint32_t v1, v2; asm volatile(\
+#define read_cntvoff()          ({ unsigned int v1, v2; asm volatile(\
                                 " mrrc     p15, 4, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
+                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
 
 #define write_cntvoff(val)    asm volatile(\
                               " mcrr     p15, 4, %0, %1, c14\n\t" \
